refactor: all exception in one file
