--- includes/exceptions/exceptionsIncludes.hpp
@@ -1,16 +1,0 @@
-#ifndef EXCEPTIONS_INCLUDES_HPP
-# define EXCEPTIONS_INCLUDES_HPP
-
-# include "NeedMoreParamsException.hpp"
-# include "AlreadyRegisteredException.hpp"
-# include "PasswordMismatchException.hpp"
-# include "NoNicknameGivenException.hpp"
-# include "ErroneousNicknameException.hpp"
-# include "NicknameInUseException.hpp"
-# include "NickCollisionException.hpp"
-# include "NotRegisteredException.hpp"
-
-# include "ServerException.hpp"
-# include "CommandNotFoundException.hpp"
-
-#endif
